Make release notes unread badge behave better
Closes FOO-3239
flag=none

We needed better behavior with the unread badge on the
help menu, which denotes the number of unread release
notes.

Observations:
  - new release notes are likely added only very
    infrequently, so there's no need to do a timed
    repoll of the unread count except rarely
  - reading the release notes by opening the help tray
    should immediately clear the badge, since opening
    the tray displays the new release notes

To that effect, we removed earlier code that polled the
release notes faster or slower depending on how many
unread release notes there were and instead added a
callback function which allows a parent component to
ask for a way to force a change to an UnreadCount badge
(either by forcing an immediate poll or by simply
providing a new value to set it to). Now we can just
use that signal to clear the badge when the tray slides
open, and reduce the polling interval to something very
infrequent. This should help the load on that API
endpoint considerably.

Test plan:
* Have release notes working (requires local dynamodb)
* Create a new release note
* Quit and restart your browser for dev Canvas (that will
  clear the session storage and force an immediate check
  of new release notes
* The badge should show on the help menu indicating new
  (unread) release notes
* Slide out the help tray by clicking on the Help icon in
  the nav sidebar and the unread badge should immediately
  disappear
* Monitor your browser's network calls in the dev tools;
  you should hardly ever see a poll for unread release
  notes (/api/v1/release_notes/unread_count)... it will
  poll only once per hour
* Loading a new page (try clicking on the dashboard and
  then visiting an admin page) should not trigger a poll
  either as the polling interval should be preserved in
  the session storage

Change-Id: I175127073c3cabed2ce6e4c5e0b52af36f457f34
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/307023
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>