Fix cached preference for enabled_account_calendars
It was possible that the @current_user would be loaded into memory with
stale preferences that caused :enabled_account_calendars preference to
not fetch the correct value. The fix here is to reload the @current_user
in the case that we are fetching the value. We do this in other locations
like the conversations_controller.rb and users_controller.rb to ensure
values are correct.

fixes LS-3660
flag=none

test plan:
- Go to Account Calendars and enable the calendar for at least one
account/sub-account.
- Find a user enrolled in a course in that account/sub-account.
- As that user, go to the Calendar page.
- Click on the + under "Other Calendars" then subscribe to an account
calendar.
- Refresh the page
- Verify the account calendar is listed under "Other Calendars"
- Refresh a few more times and make sure it's still there
- Remove the calendar from "Other Calendars" by clicking the X
- Refresh the page
- Verify the account calendar is not listed under "Other Calendars"
- Refresh a few more times and make sure it's still not there

Change-Id: I9726b2035b136c2cc5330c66103586b1ad961245
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/308151
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>