Fix loading order of locale polyfills
Refs FOO-2929
flag=none

There are not a lot of locales that need to be
polyfilled (native browser support exists in almost
all cases), so it's no surprise that it was not
properly loading some of the locale data, due to
us not checking for prerequisite polyfills and also
to not loading polyfills in the required order.

See https://formatjs.io/docs/polyfills/ ... there's
a pretty clear tree of dependencies.

This commit changes the order of loading (if needed)
of the polyfills and also brings in a couple of
polyfills that I didn't even realize were required
for what we are loading. Also I removed a not-longer-
needed directory in /packages that used to handle the
locale polyfills before it was moved to the "engine"
stuff; also now the dependencies are in the correct
package.json.

Test plan:
* re-run the test plan for g/294705
* If you are an expert at how dates should look in
  Welsh or Armenian, make sure they're really being
  done correctly now?

Change-Id: Ie65ec591b9c9c063a093dec8a921759354abc37d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/310469
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>