Allow Chrome OS LTS
Chrome OS has a long-term support (LTS) channel which is updated every
6 months. Some schooles have chosen to use this channel for their Chrome
OS updates. Unfortunately, the LTS distinction is not reflected in the
User-Agent string.

This change checks if the browser is Chrome running on Chrome OS and
compares it to the minimum version specified in the configuration file.

This does mean means that slightly older versions of Chrome on non-LTS
Chrome OS will also be be allowed to run Canvas without warnings, but
that's better than not allowing LTS versions to be supported at all.

flag=none
Closes FOO-3343

Test Plan:
  - Visit a page in Canvas with a browser that is running an LTS version
  of Chrome OS that is older than the latest support Chrome.
  - You can spoof this in Chrome dev tools by changing the User-Agent
     under "... > More Tools > Network Conditions".

Change-Id: Ic9cadd20a6170f144103249daba6d729ba41e47c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/311263
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>