Make recent feedback url work for observers
The url to access submissions from the dashboard's recent feedback
sidebar was generating the subissions url with the observer's user
id, which does not work because the observer doesn't have a submission
in the course. Instead, construct the url with the id of the student
whom the observer is currently observing.

closes LS-3934
flag = none

Test plan:
 - Create an assignment in a course with a due date and online
   submission type
 - As a student, submit the assignment
 - As a teacher, leave a submission comment
 - As an observer (observing the student), open the dashboard and
   click the link to the comment in the recent feedback sidebar
 - Expect to be taken to the student's submission and see the comment

Change-Id: I52c16f4fc2efc885e4adc298b60a1140ac191b1d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/316066
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>