Catch newly added module items while disabling for bulk publish
The reason newly added item's pub button didn't get disabled is
because it's data-* attributes are all unset. At first I thought
it made sense to get those set, but in the current code this
is done by the erb template and not in the UI, so I decided
not to touch that logic. The fix, passing the HTML element
to the updateModuleItemPublishedState instead of the item id
(which is missing) is actually better logic there than
before, and solves this problem.

closes LS-3964
flag=module_publish_menu

test plan:
  - in a course have a couple modules with items
  - from the modules page, add an item to one of the modules
    bonus points for adding an item that's in a different module
  - pub/unpub with items the module you added the item to
  > expect the newly added item's pub/unpub button to be disabled
  > expect the item in the other module to be disabled too

Change-Id: Ic4eafd60c2acc66ee62bf0e0ea31acfd8be73f77
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/316918
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>