Add ability to update event series in calendar
closes LF-496
closes LF-486

flag=calendar_series

test plan 1:
- Navigate to calendar page.
- Create a new event changing the frequency
to anything except "not repeat" or "custom".
- Click submit.
- Select the any occurrence of the series EXCEPT first.
- Click Edit.
- (Optional) Select another frequency.
- Click Submit.
> Verify that the modal only shows "this event" and
"this and following" options.

test plan 2:
- Navigate to calendar page.
- Create a new event changing the frequency
to anything except "not repeat" or "custom".
- Click submit.
- Select the first occurrence of the series.
- Click Edit.
- (Optional) Select another frequency.
- Click Submit.
> Verify that the modal only shows "all",
"this event" and "this and following" options.

test plan 3:
- Navigate to calendar page.
- Create a new event changing the frequency
to "not repeat".
- Click submit.
- Select the first occurrence of the series.
- Click Edit.
- (Optional) Select another frequency.
- Click Submit.
> Verify that the options modal is not showed
since we have only on event.
> Verify that when editing this events again
the modal will be showed this time.

On all test plans:
> Verify that after confirming the update,
the calendar is properly updated.
> Verify that after confirming the update and
refreshing the page the changes are
persisted.
> Verify all these scenarios selecting
"More Options" (Page), on submitting we
should see the modal and act the same way.

Note: Check that when creating/editing an
event with custom frequency are properly
created, loaded and edited.

Change-Id: Ib3c44bae7f0f3cefa5ac91a7f61cb2452035c325
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/323082
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Juan Chavez <juan.chavez@instructure.com>