Fix errant popup menu display in Permissions
Refs FOO-3190
Fixes FOO-3858
flag=none

At InstUI 8, the popup menu that appears next to a permission
button when it is clicked on loses its "trigger" relationship
with the button and instead "pops up" at the bottom of the page.

This is due to the wrapping of the permission button in a theme
override `<InstUISettingsProvider>` component, which I guess
breaks the "trigger" relationship with the parent menu. Only one
theme variable is being overridden here, though, so it's easy
enough to do without InstUISettingsProvider and simply provide
a `themeOverride` prop to the button itself.

Also, at InstUI 8 we were now seeing error messages due to
incorrect props being sent to <Menu> ... we had a Menu.Group
with a `selected` prop to indicate where to display the check
marks on the menu, but in that case the group needs to have an
`onSelect` prop for a callback when the menu is changed. This
had recently been done by individual `onClick` props on the
menu items themselves, but this is not really the right way to
go about managing a menu. And after changing out all the onClick
props for a single onSelect on the group, the code winds up
being a lot cleaner anyway.

Test plan:
* Visit the permissions page (/accounts/:id/permissions)
* Clicking on the red or green icons in the table should now
  bring up a menu right next to the button
* Exercise the various options on the menu... they should still
  cause the same effects on the permission as previously.

Change-Id: I91c7e96fec95127d03b69b3b75e3564d1420d648
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/328056
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>