do not load speedgrader with only test student
When loading the speedgrader in with a section filter, we were checking
if any students are in that section and also assigned to the assignment
and by default the test student is in every section so it would always
load the speedgrader. This change makes it so we check if we are only
loading the  test student to remove the filter and show all sections
which was already the expected behavior.

flag=enhanced_gradebook_filters
closes EVAL-3572

Test Plan:
- Create a course with a few students
- Create 2 sections and add some students to each
- create an assignment only assigned to one section
- as a teacher filter the gradebook by the section that the assignment
  is not assigned to
- load the assignment in speedgrader
- ensure that an alert is shown reading: "Could not find any students in
 that section, falling back to showing all sections."
- after clicking OK ensure all assigned students are shown in speed
 grader

Change-Id: Ibbdd04c552578749bb64ae102fd5b73c83aa3acc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/330417
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Derek Williams <derek.williams@instructure.com>
Reviewed-by: Samuel Lee <samuel.lee@instructure.com>
QA-Review: Derek Williams <derek.williams@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>